pyproject.toml
[project]
name = "pl-knowledge-engine"
version = "0.1.0"
description = "Multimodal knowledge engine for Premier League analysis — Aston Villa vs Liverpool FC (2025-26)"
requires-python = ">=3.11"
dependencies = [
    # Core
    "python-dotenv>=1.0.0",
    "pydantic-settings>=2.0.0",
    "structlog>=24.0.0",

    # Data manipulation
    "pandas>=2.1.0",
    "requests>=2.31.0",
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",

    # YouTube & media
    "youtube-transcript-api>=1.0.0",
    "yt-dlp>=2024.0.0",
    "google-api-python-client>=2.100.0",

    # Audio processing
    "pydub>=0.25.1",

    # AI & embeddings
    "google-genai>=1.0.0",
    "anthropic>=0.40.0",

    # Vector database
    "chromadb>=0.5.0",

    # Knowledge graph
    "neo4j>=5.14.0",

    # Testing
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
]

[project.optional-dependencies]
dev = [
    "ipykernel>=6.0.0",
    "jupyter>=1.0.0",
    "ruff>=0.5.0",
    "mypy>=1.8.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
